Fix failures on stable Rust
`pub(restricted)` is only on the beta channel right now.
